@@ -5,7 +5,7 @@ group = 'com.github.cookpad'
 version = '0.0.1'
 
 android {
-    compileSdkVersion 27
+    compileSdkVersion 28
 
     testOptions {
         unitTests {
@@ -15,11 +15,11 @@ android {
 
     defaultConfig {
         minSdkVersion 16
-        targetSdkVersion 27
+        targetSdkVersion 28
         versionCode 1
         versionName "1.0"
 
-        testInstrumentationRunner "android.support.test.runner.AndroidJUnitRunner"
+        testInstrumentationRunner "androidx.test.runner.AndroidJUnitRunner"
     }
 
     buildTypes {
@@ -33,19 +33,19 @@ android {
 
 dependencies {
     implementation fileTree(dir: 'libs', include: ['*.jar'])
-    implementation "android.arch.lifecycle:extensions:1.1.1"
-    kaptAndroidTest "android.arch.lifecycle:compiler:1.1.1"
+    implementation "androidx.lifecycle:lifecycle-extensions:2.0.0"
+    kaptAndroidTest "androidx.lifecycle:lifecycle-compiler:2.0.0"
 
-    implementation 'com.android.support:appcompat-v7:27.1.1'
+    implementation 'androidx.appcompat:appcompat:1.0.2'
     implementation "org.jetbrains.kotlin:kotlin-stdlib:$kotlin_version"
     implementation "io.reactivex.rxjava2:rxjava:2.1.12"
 
     testImplementation 'junit:junit:4.12'
-    testImplementation 'com.nhaarman:mockito-kotlin:1.5.0'
-    testImplementation 'com.google.truth:truth:0.39'
+    testImplementation 'com.nhaarman:mockito-kotlin:1.6.0'
+    testImplementation 'com.google.truth:truth:0.40'
     testImplementation "org.robolectric:robolectric:3.5.1"
-    androidTestImplementation 'com.android.support.test:runner:1.0.1'
-    androidTestImplementation 'com.android.support.test.espresso:espresso-core:3.0.1'
+    androidTestImplementation 'androidx.test:runner:1.1.0'
+    androidTestImplementation 'androidx.test.espresso:espresso-core:3.1.0'
 }
 
 repositories {
