@@ -1,9 +1,9 @@
 package org.cookpad.app_test.home
 
 import android.os.Bundle
-import android.support.v4.app.Fragment
-import android.support.v4.app.FragmentPagerAdapter
-import android.support.v7.app.AppCompatActivity
+import androidx.appcompat.app.AppCompatActivity
+import androidx.fragment.app.Fragment
+import androidx.fragment.app.FragmentPagerAdapter
 import io.reactivex.subjects.PublishSubject
 import kotlinx.android.synthetic.main.activity_recipe_list.*
 import org.cookpad.app_test.R
